% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3.R
\docType{data}
\name{gadmISO3}
\alias{gadmISO3}
\title{Country iso 3 codes for UI}
\format{
A data frame with all countries and their current ISO3 codes as used by GADM
}
\source{
from the raster package
}
\usage{
gadmISO3
}
\description{
ISO3 codes and country names for GADM data
}
\keyword{datasets}
