% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runSpatSampleApp}
\alias{runSpatSampleApp}
\title{Start the Survey Solutions Spatial Sampling Application}
\usage{
runSpatSampleApp(
  launch.browser = T,
  mapwidget.option = c("mapdeck", "leaflet"),
  google_ai_score_url =
    "https://storage.googleapis.com/open-buildings-data/v1/score_thresholds_s2_level_4.csv",
  google_ai_score_map_url = "https://sites.research.google/open-buildings/tiles.geojson",
  bufferForSuSoBounds = 5,
  pointsLimit = 1e+05
)
}
\arguments{
\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only. The value of this parameter can also be a
function to call with the application's URL.}

\item{mapwidget.option}{Selection of map at start-up}

\item{google_ai_score_url}{URL for google AI scores, unlikely to change!}

\item{google_ai_score_map_url}{Url for base map to google AI, unlikely to change!}

\item{bufferForSuSoBounds}{Buffere for Survey Solutions geo-fencing}

\item{pointsLimit}{Limit for points data to be aggregated to area}
}
\description{
A wrappter function to start the application. Please make sure you have read the
documentation on how to use the app under \url{https://datanalytics.worldbank.org/SpatialSamplingManual/}
}
\details{
This application is part of the large set of tools, to facilitate survey implementation with
\href{https://docs.mysurvey.solutions/}{Survey Solutions}. Enumeration Areas, sampled for example
with the JDC-Survey Solutions Spatial Sampling application, can be used to enumerate all buildings
visible within the boundaries on a Google map. The enumerate buildings can then be used for the
second stage sampling frame and to draw the survey units within the cluster from it.
}
\examples{
\dontrun{
# Start app in the current working directory
runApp()

# Start app in a subdirectory called myapp
runApp("myapp")
}

## Only run this example in interactive R sessions
if (interactive()) {
  options(device.ask.default = FALSE)

  # Apps can be run without a server.r and ui.r file
  runApp(list(
    ui = bootstrapPage(
      numericInput('n', 'Number of obs', 100),
      plotOutput('plot')
    ),
    server = function(input, output) {
      output$plot <- renderPlot({ hist(runif(input$n)) })
    }
  ))


  # Running a Shiny app object
  app <- shinyApp(
    ui = bootstrapPage(
      numericInput('n', 'Number of obs', 100),
      plotOutput('plot')
    ),
    server = function(input, output) {
      output$plot <- renderPlot({ hist(runif(input$n)) })
    }
  )
  runApp(app)
}
}
