% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runappserver.R
\name{runSpatSampleAppServer}
\alias{runSpatSampleAppServer}
\title{Start the Survey Solutions Spatial Sampling Application on a Shiny Server}
\usage{
runSpatSampleAppServer(
  mapwidget.option = c("leaflet", "mapdeck"),
  google_ai_score_url =
    "https://storage.googleapis.com/open-buildings-data/v1/score_thresholds_s2_level_4.csv",
  google_ai_score_map_url = "https://sites.research.google/open-buildings/tiles.geojson",
  bufferForSuSoBounds = 5,
  pointsLimit = 1e+05
)
}
\arguments{
\item{mapwidget.option}{Selection of map at start-up, mapdeck or leaflet, defaults to leaflet if NULL}

\item{google_ai_score_url}{URL for google AI scores, unlikely to change!}

\item{google_ai_score_map_url}{Url for base map to google AI, unlikely to change!}

\item{bufferForSuSoBounds}{Buffer for Survey Solutions
\href{https://docs.mysurvey.solutions/syntax-guide/questions/syntax-guide-gps-questions/}{Geofencing}}

\item{pointsLimit}{Limit for points data to be aggregated to area}
}
\description{
A wrappter function to start the application on the SERVER. Please make sure you have read the
documentation on how to use the app under \url{https://datanalytics.worldbank.org/SpatialSamplingManual/}
}
\details{
This application is part of the large set of tools, to facilitate survey implementation with
\href{https://docs.mysurvey.solutions/}{Survey Solutions}. Enumeration Areas, sampled for example
with the JDC-Survey Solutions Spatial Sampling application, can be used to enumerate all buildings
visible within the boundaries on a Google map. The enumerate buildings can then be used for the
second stage sampling frame and to draw the survey units within the cluster from it.
}
