% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susospatsample-package.R
\docType{package}
\name{susospatsample-package}
\alias{susospatsample}
\alias{susospatsample-package}
\title{susospatsample: Shiny Application for Spatial Sample and Survey Solutions resource creation}
\description{
A comprehensive Shiny application to process and sample from spatial resources. Spatial area frames can be created either through the creation of a regular grid, the use of gridded population data or from Enumeration areas boundaries. Besides sampling from an area frame (i.e. Enumeration Areas), the application also allows for sampling from geo-referenced points (i.e. Buildings). When the sampling is completed resources for data collection with the Survey Solutions CAPI system can be created, and directly assigned to the server.
}
\author{
\strong{Maintainer}: Michael Wild \email{mwild@worldbank.org}

}
\keyword{internal}
